server/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
pillow==11.2.1
python-pptx==1.0.2
typing_extensions==4.14.0
Werkzeug==3.1.3
XlsxWriter==3.2.3
